memory/tencentdb: split adapter tests
